[kdbus] Add new library API for common tasks on dbus/kdbus
For kdbus purpose we need introduce two new bus types
called "user" and "machine". Will not be possible to
make calls to the 'org.freedesktop.DBus' destination on
the new bus types, thus this patch introduce new API
for tasks such as "get Unix user ID of the process connected
to the server", "returns a list of all currently-owned names
on the bus" and so on.

Change-Id: Ib9ba65346290d597341d4157b09cb45bd3d1151d